Remove preflight stage login window assumption
Removes the code that checks for the login window when executing the
preflight stage of installapplications.py, and updates the README to
reflect this change.

Previously, if the preflight stage were run at the loginwindow,
installapplications assumed it was running on a new machine. While this
assumption may hold for certain workflows, certain MDM setups or
workflows may include sending the InstallApplication MDM command to a
device at the loginwindow, which would automatically skip the preflight
and it would sit there until next user logged in. This may result in
unintended DEPNotify runs and/or logouts/restarts for users with
previously deployed and configured machines. This assumption has been
removed to instead allow the preflight script to handle this logic.