Move the main loop that processes itemsToProcess (after parsing CLI args) and construction of stages dict out to a function.
The function takes only the itemsToProcess list that is
constructed by parsing the args.item inputs as well as the base-url.

Within the function the code is essentially unchanged, and the
function returns the completed stages dictionary.

The only other change is to remove the branch "if args.output" when
constructing the savePath...args.output is required, so it will always
be set (also, the alternative 'rootdir' appears to not be defined).