generatejson - cast retries and retrywait to ints
By default, argparse is using str for the retries and retrywait. This change forces a cast to int so that we don't end up with strings being read into IAs itself (throws a TypeError when it hits the retries code path).

Ideally I'd have argparse validate this and just store it as an int, but with using nargs I can't find an elegant way to do that without a lot of refactoring and I don't think that's worth it.